refactor: change types to interfaces + renaming files
